requirements.txt
scapy==2.5.0
python-dotenv==1.0.0
pandas==2.1.0
jinja2==3.1.2
psutil==5.9.5
